#50 refactor: time-control-03 の import 元を hooks/event に更新
@/hooks/use-event-listener・@/hooks/use-event-dispatcher の廃止に追従
