+++ src/prototypes/time-control/time-control-03/_events/_hooks/use-time-control-03-event-dispatcher.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react'
 
-import { useEventDispatcher } from '@/hooks/use-event-dispatcher'
+import { useEventDispatcher } from '@/hooks/event'
 
 import { useScopeEventTarget } from '../_contexts/scope-event-context'
 import { TimeControl03EventMap } from '../index.types'
