+++ src/prototypes/time-control/time-control-03/_events/_hooks/use-time-control-03-event-pending.ts
@@ -1,4 +1,4 @@
-import { useEventPending } from '@/hooks/use-event-listener'
+import { useEventPending } from '@/hooks/event'
 
 import { useScopeEventTarget } from '../_contexts/scope-event-context'
 import { TimeControl03EventMap } from '../index.types'
